# apps/server/internal/api/openapi.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package api

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// CreateStaticRequest defines model for CreateStaticRequest.
type CreateStaticRequest struct {
	Callsign *string `json:"callsign,omitempty"`
	Imo      int     `json:"imo"`
	Mmsi     int     `json:"mmsi"`
	MsgType  int     `json:"msg_type"`
	ShipType *string `json:"ship_type,omitempty"`
	Shipname string  `json:"shipname"`
}

// DynamicHistoryItem defines model for DynamicHistoryItem.
type DynamicHistoryItem struct {
	Course     *float32   `json:"course,omitempty"`
	Latitude   *float32   `json:"latitude,omitempty"`
	Longitude  *float32   `json:"longitude,omitempty"`
	SpeedKnots *float32   `json:"speed_knots,omitempty"`
	Status     *string    `json:"status,omitempty"`
	Timestamp  *time.Time `json:"timestamp,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error *string `json:"error,omitempty"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Database *string `json:"database,omitempty"`
	Service  *string `json:"service,omitempty"`
	Status   string  `json:"status"`
}

// ShipDetailsResponse defines model for ShipDetailsResponse.
type ShipDetailsResponse struct {
	Callsign               *string               `json:"callsign,omitempty"`
	Imo                    *int                  `json:"imo,omitempty"`
	Mmsi                   int                   `json:"mmsi"`
	PositionHistory        *[]DynamicHistoryItem `json:"position_history,omitempty"`
	ShipType               *string               `json:"ship_type,omitempty"`
	Shipname               string                `json:"shipname"`
	TotalPositionsRecorded *int                  `json:"total_positions_recorded,omitempty"`
}

// GetShipByMMSIParams defines parameters for GetShipByMMSI.
type GetShipByMMSIParams struct {
	Limit  *int `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// GetShipByIMOParams defines parameters for GetShipByIMO.
type GetShipByIMOParams struct {
	Limit  *int `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// CreateStaticAISJSONRequestBody defines body for CreateStaticAIS for application/json ContentType.
type CreateStaticAISJSONRequestBody = CreateStaticRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Consultar buque e historial de telemetría por MMSI
	// (GET /api/v1/ship/mmsi/{mmsi})
	GetShipByMMSI(w http.ResponseWriter, r *http.Request, mmsi int, params GetShipByMMSIParams)
	// Consultar buque e historial de telemetría por IMO
	// (GET /api/v1/ship/{imo})
	GetShipByIMO(w http.ResponseWriter, r *http.Request, imo int, params GetShipByIMOParams)
	// Registrar o actualizar información estática del buque
	// (POST /api/v1/static)
	CreateStaticAIS(w http.ResponseWriter, r *http.Request)
	// Health check profundo
	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetShipByMMSI operation middleware
func (siw *ServerInterfaceWrapper) GetShipByMMSI(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "mmsi" -------------
	var mmsi int

	err = runtime.BindStyledParameterWithOptions("simple", "mmsi", r.PathValue("mmsi"), &mmsi, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "mmsi", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetShipByMMSIParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "offset", r.URL.Query(), &params.Offset, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "offset"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetShipByMMSI(w, r, mmsi, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetShipByIMO operation middleware
func (siw *ServerInterfaceWrapper) GetShipByIMO(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "imo" -------------
	var imo int

	err = runtime.BindStyledParameterWithOptions("simple", "imo", r.PathValue("imo"), &imo, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "imo", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetShipByIMOParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "offset", r.URL.Query(), &params.Offset, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "offset"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetShipByIMO(w, r, imo, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateStaticAIS operation middleware
func (siw *ServerInterfaceWrapper) CreateStaticAIS(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateStaticAIS(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/ship/mmsi/{mmsi}", wrapper.GetShipByMMSI)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/ship/{imo}", wrapper.GetShipByIMO)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/static", wrapper.CreateStaticAIS)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/health", wrapper.GetHealth)

	return m
}